Convert output to simplified tree
The final output should resemeble something like a VDOM structure.

These changes recursively strip away the information from the parse and
turn it into an easier output tree of `[ type, attrs, children ]` where
`type` is a pair of `[ major type, minor type ]` and `major type` could
be `block` or `tag` (for an HTML tag).

This is just one path towards forming a language with which to
communicate with plugins. The idea is that this format is trivial to
abstract with functions.