Decode entities in text nodes
Note that this should involve a build stage so that the parser itself
doesn't look so ugly. I have included the body of `he` into the global
function area of the parser to give us `decode()`. `he` is a robust
library for encoding and decoding HTML entities.

The purpose of this change is to demonstrate how this is possible to
keep the character and entity encoding fully abstracted inasmuch as it
concerns the block renderers, editors, and core.